fix: make it bundle
